Bug Fix: fix backwards compatibility issue
Previously, when creating a signature with a payload of `None`, `""` or
`{}` would all result in the same signature. After the changes in 3.0
this is no longer true. Each of those 3 scenarios result in a different
signature. This is a problem for the other implementations of the
request signer (Perl, C#). With this change signatures with no payload
(whether it is `""`, `{}` or `None`) will all be the same (which works
like the `""` version in the broken 3.0 version - which is what happened
before 3.0).